requirements.txt
streamlit>=1.20.0
GitPython>=3.1.31
pandas>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
